+++ package.json
@@ -5,7 +5,7 @@
   "main": "server/app.js",
   "scripts": {
     "start": "nodemon --exec babel-node server/app.js",
-    "test": "mocha --require @babel/register server/v2/test/*.js --exit || true",
+    "test": "mocha --require @babel/register server/test/*.js --exit || true",
     "build": "babel server --out-dir dist",
     "serve": "node dist/app.js",
     "heroku-postbuild": "npm install",
