+++ server/test/message.test.js
@@ -1,6 +1,6 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import server from '../../app';
+import server from '../app';
 
 const should = chai.should();
