--- server/v1/model/contact.js
@@ -1,10 +1,0 @@
-export default class Contacts {
-  constructor() {
-    this.id = null;
-    this.firstName = null;
-    this.lastName = null;
-    this.email = null;
-    this.contactOwner = null;
-    this.createdAt = new Date();
-  }
-}
