--- server/v1/model/draft.js
@@ -1,9 +1,0 @@
-export default class Drafts {
-  constructor() {
-    this.id = null;
-    this.senderId = null;
-    this.messageId = null;
-    this.receiverId = null;
-    this.createdOn = new Date();
-  }
-}
