--- server/v1/model/inbox.js
@@ -1,10 +1,0 @@
-export default class Inboxs {
-  constructor() {
-    this.id = null;
-    this.receiverId = null;
-    this.senderId = null;
-    this.messageId = null;
-    this.read = false;
-    this.createdOn = new Date();
-  }
-}
