--- server/v1/model/sent.js
@@ -1,9 +1,0 @@
-export default class Sents {
-  constructor() {
-    this.id = null;
-    this.senderId = null;
-    this.receiverId = null;
-    this.messageId = null;
-    this.createdOn = new Date();
-  }
-}
