--- server/v1/model/user.js
@@ -1,11 +1,0 @@
-export default class Users {
-  constructor() {
-    this.id = null;
-    this.firstName = null;
-    this.lastName = null;
-    this.email = null;
-    this.password = null;
-    this.createdAt = new Date();
-    this.updateAt = new Date();
-  }
-}
