--- server/v1/routes/api.js
@@ -1,7 +1,0 @@
-import userRoute from './user';
-import messageRoute from './message';
-
-module.exports = {
-  userRoute,
-  messageRoute,
-};
